[FIX] Admin: Fixes around changing label cases

+++ app/src/Application/DeskPRO/EntityRepository/LabelDef.php
@@ -191,7 +191,7 @@ class LabelDef extends AbstractEntityRepository
 
 		$ret = array();
 		$res = $db->executeQuery(sprintf(
-			'SELECT LOWER(label) as label FROM %s WHERE label_type = :type', $this->getTableName()
+			'SELECT label FROM %s WHERE label_type = :type', $this->getTableName()
 		), array('type' => $type));
 
 		while ($row = $res->fetchColumn(0)) {
@@ -247,10 +247,10 @@ class LabelDef extends AbstractEntityRepository
 	public function getDefinition($type, $label)
 	{
 		return $this->getEntityManager()->createQuery(
-			'SELECT d FROM DeskPRO:LabelDef d WHERE d.label_type = :type AND LOWER(d.label) = :label'
+			'SELECT d FROM DeskPRO:LabelDef d WHERE d.label_type = :type AND d.label = :label'
 		)->setParameters(array(
 			'type' => $type,
-			'label' => strtolower(trim($label)),
+			'label' => trim($label),
 		))->getOneOrNullResult();
 	}
 
@@ -263,7 +263,7 @@ class LabelDef extends AbstractEntityRepository
 		$counts = $this->countDefUsages();
 
 		foreach ($definitions as &$def) {
-			$label = strtolower($def['label']);
+			$label = $def['label'];
 			$def['total'] = isset($counts[$def['label_type']][$label]) ? $counts[$def['label_type']][$label] : 0;
 		}
 
@@ -275,8 +275,17 @@ class LabelDef extends AbstractEntityRepository
 	 */
 	public function updateDefinitionUsages(\Application\DeskPRO\Entity\LabelDef $definition)
 	{
+		// Need to run an update to change cases because table is case-insensitive
+		$this->_em->getConnection()->executeUpdate("
+			UPDATE IGNORE label_defs SET label = ? WHERE label = ?
+		", array($definition->label, $definition->label));
+
+		$this->_em->getConnection()->executeUpdate(sprintf("
+			UPDATE IGNORE %s SET label = ? WHERE label = ?
+		", self::$types[$definition['label_type']]['table']), array($definition->label, $definition->label));
+
 		$counts = $this->countDefUsages(array($definition['label_type']));
-		$label = strtolower($definition['label']);
+		$label = $definition['label'];
 		$definition['total'] = isset($counts[$definition['label_type']][$label])
 			? $counts[$definition['label_type']][$label]
 			: 0;
@@ -293,7 +302,7 @@ class LabelDef extends AbstractEntityRepository
 		try {
 			$this->getEntityManager()->getConnection()->executeUpdate(
 				sprintf('DELETE FROM %s WHERE label = ?', self::$types[$definition['label_type']]['table']),
-				array(strtolower($definition['label']))
+				array($definition['label'])
 			);
 			$this->getEntityManager()->remove($definition);
 			$this->getEntityManager()->flush();
@@ -310,15 +319,15 @@ class LabelDef extends AbstractEntityRepository
 		$this->getEntityManager()->createQuery('
 			UPDATE DeskPRO:LabelDef l
 			SET l.color = :color
-			WHERE LOWER(l.label) = :label AND l.label_type = :label_type
-		')->execute(array('label_type' => $type, 'label' => strtolower($label), 'color' => $color));
+			WHERE l.label = :label AND l.label_type = :label_type
+		')->execute(array('label_type' => $type, 'label' => $label, 'color' => $color));
 	}
 
 	public function getColorForLabel($label)
 	{
 		$q = $this->getEntityManager()->createQuery('
-			SELECT d.color FROM DeskPRO:LabelDef d WHERE LOWER(d.label) = :label
-		')->setMaxResults(1)->setParameters(array('label' => strtolower($label)));
+			SELECT d.color FROM DeskPRO:LabelDef d WHERE d.label = :label
+		')->setMaxResults(1)->setParameters(array('label' => $label));
 
 		$res = $q->getScalarResult();
 
@@ -346,7 +355,7 @@ class LabelDef extends AbstractEntityRepository
 			if ($k > 0) {
 				$query .= "\n UNION ";
 			}
-			$query .= 'SELECT "'.$t.'" as label_type, COUNT(*) AS count, LOWER(label) as label FROM ' . $info['table'] . ' GROUP BY label';
+			$query .= 'SELECT "'.$t.'" as label_type, COUNT(*) AS count, label as label FROM ' . $info['table'] . ' GROUP BY label';
 		}
 
 		$count_res = $this->getEntityManager()->getConnection()->fetchAll($query);
@@ -425,12 +434,18 @@ class LabelDef extends AbstractEntityRepository
 
 				$this->getEntityManager()->getConnection()->executeUpdate(
 					'UPDATE IGNORE ' . $table . ' SET label = ? WHERE label = ?',
-					array($new_label, strtolower($old_label))
+					array($new_label, $old_label)
 				);
-				$this->getEntityManager()->getConnection()->executeUpdate(
-					'DELETE FROM ' . $table . ' WHERE label = ?',
-					array(strtolower($old_label))
-				);
+
+				// Same one -- we are just changing case
+				if ($def_new && $def_old && $def_old === $def_old) {
+					$def_new->label = $new_label;
+				} else {
+					$this->getEntityManager()->getConnection()->executeUpdate(
+						'DELETE FROM ' . $table . ' WHERE label = ?',
+						array($old_label)
+					);
+				}
 
 				if ($def_old && $def_new) {
 					$this->getEntityManager()->remove($def_old);
